requirements.txt
beautifulsoup4==4.12.2
Django==4.1.9
djangorestframework==3.14.0
requests==2.27.1
tqdm==4.64.1
django-cors-headers